+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+* Remove `inline` attributes on trait method side. It can avoid [`clippy::inline_fn_without_body`](https://rust-lang.github.io/rust-clippy/master/index.html#inline_fn_without_body) by this.
+
 # 0.1.1 - 2019-02-21
 
 * Fix an error related to generics.